Application SmartParking/Backend/requirements.txt
dj_database_url
django-cors-headers
djangorestframework
mysqlclient
opencv-python
ultralytics
easyocr